packages/ai-orchestrator/pyproject.toml
[project]
name = "ai-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Kelby Madal-Hellmuth", email = "kelby0320@gmail.com" }]
requires-python = ">=3.14"
dependencies = [
    "ai-core",
    "ai-infra",
    "anyio>=4.11.0",
    "fastapi>=0.121.2",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "httpx>=0.28.1",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-grpc",
    "opentelemetry-exporter-otlp-proto-grpc",
    "opentelemetry-instrumentation-logging",
    "python-json-logger",
    "uvicorn>=0.38.0",
]

[project.scripts]
ai-orchestrator = "ai_orchestrator:main"

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
ai-core = { workspace = true }
ai-infra = { workspace = true }
